Parallelized `scripts/build.js`
