+++ package.json
@@ -39,6 +39,7 @@
     "lerna": "2.0.0-rc.4",
     "micromatch": "^2.3.11",
     "mkdirp": "^0.5.1",
+    "pify": "^3.0.0",
     "prettier": "1.3.1",
     "progress": "^1.1.8",
     "react": "15.4.2",
@@ -49,7 +50,8 @@
     "rimraf": "^2.5.4",
     "slash": "^1.0.0",
     "strip-ansi": "^3.0.1",
-    "typescript": "^2.2.2"
+    "typescript": "^2.2.2",
+    "worker-farm": "^1.3.1"
   },
   "scripts": {
     "build-clean": "rm -rf ./packages/*/build ./packages/*/build-es5",
